+++ src/test/groovy/groovylicioustest/AutoRegisterTest.groovy
@@ -1,6 +1,6 @@
 package groovylicioustest
 
-import ga.ozli.minecraftmods.groovylicious.transform.AutoRegister
+import ga.ozli.minecraftmods.groovylicious.transform.Registroid
 import groovy.transform.CompileStatic
 import groovy.transform.stc.POJO
 import net.minecraft.core.Registry
@@ -9,14 +9,13 @@ import net.minecraft.sounds.SoundEvent
 import net.minecraft.world.item.BlockItem
 import net.minecraft.world.item.Item
 import net.minecraft.world.level.block.Block
-import net.minecraft.world.level.block.entity.BlockEntityType
 import net.minecraft.world.level.block.state.BlockBehaviour
 import net.minecraft.world.level.material.Material
 import net.minecraftforge.registries.ForgeRegistries
 
 @POJO
 @CompileStatic
-@AutoRegister({ [Registry.ITEM, ForgeRegistries.BLOCKS, Registry.SOUND_EVENT_REGISTRY] })
+@Registroid({ [Registry.ITEM, ForgeRegistries.BLOCKS, Registry.SOUND_EVENT_REGISTRY] })
 class AutoRegisterTest {
     static final SoundEvent MY_SOUND = new SoundEvent(new ResourceLocation('groovylicioustest', 'my_sound'))
     static final Block MY_BLOCK = new Block(BlockBehaviour.Properties.of(Material.DIRT))
